backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
openai>=1.30.0
python-dotenv==1.0.0
pydantic>=2.0.0
httpx>=0.24.0
redis==5.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
asyncio
